pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cbc-omega-squared"
version = "1.0.0"
description = "CBC-Ω²: Biometric-Calibrated Cognitive Task Delegation System"
readme = "README.md"
license = {text = "MIT"}
authors = [{name="CBC-Ω² Team"}]
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.110.0",
  "uvicorn[standard]==0.29.0",
  "pydantic==2.7.0",
  "numpy==1.26.4",
  "scipy==1.13.0",
  "scikit-learn==1.4.2",
  "pandas==2.2.1",
  "prometheus-client==0.20.0",
  "python-jose[cryptography]==3.3.0",
  "httpx==0.27.0",
  "pyyaml==6.0.1",
  "pyarrow==16.1.0",
]

[project.optional-dependencies]
memory = [
  "neo4j==5.19.0",
  "sentence-transformers==2.6.1",
  "torch>=2.1.2",
]
qa = [
  "pytest==8.2.0",
  "pytest-asyncio==0.23.7",
  "pytest-cov==5.0.0",
  "ruff==0.6.8",
  "mypy==1.10.0",
  "coverage==7.6.1",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --cov=core --cov=infrastructure --cov-report=term-missing"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
lint.select = ["E", "F", "I", "B", "UP", "N", "C90"]
lint.ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unreachable = true
ignore_missing_imports = true
